+++ src/services/share.ts
@@ -1,6 +1,6 @@
 import type { PartialDeep } from 'type-fest';
 
-import { LOBE_URL_IMPORT_NAME } from '@/const/url';
+import { HERMES_URL_IMPORT_NAME } from '@/const/url';
 import { UserSettings } from '@/types/user/settings';
 
 class ShareService {
@@ -10,7 +10,7 @@ class ShareService {
    * @returns The share settings URL.
    */
   public createShareSettingsUrl = (settings: PartialDeep<UserSettings>) => {
-    return `/?${LOBE_URL_IMPORT_NAME}=${encodeURI(JSON.stringify(settings))}`;
+    return `/?${HERMES_URL_IMPORT_NAME}=${encodeURI(JSON.stringify(settings))}`;
   };
 
   /**